equip pickaxe when going to mine

+++ src/characters/Dude.ts
@@ -292,12 +292,18 @@ export class Dude extends Component implements DialogueSource {
     }
 
     setWeapon(type: WeaponType) {
+        if (this.weapon?.getType() === type) {
+            return
+        }
         this.weapon?.delete()
         this._weapon = this.entity.addComponent(WeaponFactory.make(type, this.type))
         this.shield?.setOnBack(false) // keep em in sync
     }
 
     setShield(type: ShieldType) {
+        if (this.shield?.type === type) {
+            return
+        }
         this.shield?.delete()
         this._shield = this.entity.addComponent(ShieldFactory.make(type))
         this.weapon?.setSheathed(false) // keep em in sync
